refactor: extract RecordingSessionStopReadinessPresenter.swift from RecordingSessionPresenters.swift
Splits the `RecordingSessionStopReadinessPresenter` (~35 lines) out of
RecordingSessionPresenters.swift into its own file.

Byte-identical move of lines 118-152. Both files import Combine +
Foundation.

RecordingSessionPresenters.swift: 190 → 153 lines. Tests: 667.
